+++ src/lib/constants.js
@@ -3,7 +3,7 @@ module.exports = {
   // from scryfall. This should be changed whenever there's a new
   // field added to the stored decks. For organizational puproses,
   // the naming convention is YYYY-MM-DD__reason_for_change
-  VERSION: '2018-08-22__initial_version',
+  VERSION: '2018-08-23__add_color_identity_to_card_objects',
 
   DECK_LIST_TYPES: [
     'mainDeck',