+++ src/lib/scryfall.js
@@ -18,6 +18,7 @@ function formatCard (response, card = {}) {
   card.name = response.name
   card.typeLine = response.type_line
   card.manaCost = response.mana_cost || (response.card_faces && response.card_faces[0].mana_cost) || ''
+  card.colorIdentity = response.color_identity
   card.price = {
     usd: response.usd,
     eur: response.eur,