+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Add automatic paging of search results
 * Add keyboard shortcuts
 * Move format to main page out of settings
+* Display decks color identity based on deck list
 
 # 0.2.0
 ## Features